math_tutor_backend/pyproject.toml
[project]
name = "math-tutor-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "celery>=5.4.0",
    "channels>=4.2.0",
    "channels-redis>=4.2.1",
    "daphne>=4.1.2",
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "django-model-utils>=5.0.0",
    "django-stubs>=5.1.3",
    "djangorestframework>=3.15.2",
    "mypy-extensions>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "sympy>=1.13.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-stubs>=5.1.3",
    "factory-boy>=3.3.3",
    "ipdb>=0.13.13",
    "ipython>=8.32.0",
    "isort>=6.0.0",
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "pytest-django>=4.10.0",
    "python-lsp-server[all]>=1.12.2",
    "ruff>=0.9.6",
    "types-beautifulsoup4>=4.12.0.20250204",
    "types-requests>=2.32.0.20241016",
]


[tool.pylint]
disable = ["C0114", "C0115", "C0116"]

[tool.pylsp]
plugins = ["pylint"]

[tool.pyflakes]
max-line-length = 88

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
extend-exclude = '.*/migrations/.*'

[tool.mypy]
plugins = "mypy_django_plugin.main"
exclude = ["/migrations/"]
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_reexport = true
show_error_codes = true
strict_optional = false
warn_return_any = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
[tool.django-stubs]
DJANGO_SETTINGS_MODULE = "math_tutor.settings"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "math_tutor.settings"
python_files = "tests.py test_*.py *_tests.py"

[tool.pyright]
venvPath = "./"
venv = ".venv"
reportMissingImports = true
reportMissingTypeStubs = false
typeCheckingMode = "off"
